Add: Added directory structure for APIs
