+++ hub/views/admin/repository.py
@@ -5,7 +5,7 @@ import base64
 from django.views import View
 from django.http import JsonResponse
 
-from .models import Exporter, Release, Category
+from hub.models import Exporter, Release
 from my_settings import TOKEN
 
 api_url='https://api.github.com/repos/'
@@ -100,59 +100,14 @@ class RepositoryView(View):
         except KeyError:
             return JsonResponse({'message':'KEY_ERROR'}, status=400)
 
-class CategoryView(View):
-    def get(self, request):
-        categories=Category.objects.all()
-        data={"categories":
-            [{   
-                "category_id"  : category.id,
-                "category_name": category.name
-            } for category in categories]
-        }
-        return JsonResponse(data, status=200)
-
-class MainView(View):
-    def get(self, request):
+    def delete(self, request):
         try:
-            exporters=Exporter.objects.select_related('category', 'official').prefetch_related('release_set').order_by('id')
-            data={"exporters":
-                [
-                    {
-                        "exporter_id"    : exporter.id,
-                        "name"           : exporter.name,
-                        "logo_url"       : exporter.logo_url,
-                        "category"       : exporter.category.name,
-                        "official"       : exporter.official.name,
-                        "stars"          : exporter.stars,
-                        "repository_url" : exporter.repository_url,
-                        "description"    : exporter.description,
-                        "release"        : [{
-                            "release_version": release.version,
-                            "release_date"   : release.date,
-                            "release_url"    : release.release_url
-                        } for release in exporter.release_set.all()],
-                    }
-                for exporter in exporters]
-            }
-
-            return JsonResponse(data, status=200)
-        except Exception as e:
-            return JsonResponse({'message':f"{e}"}, status=400)
-
-class DetailView(View):
-    def get(self, request, exporter_id):
-        try:
-            readme=Exporter.objects.get(id=exporter_id).readme
-            return JsonResponse({"data":readme.decode('utf-8')}, status=200)
-
-        except Exporter.DoesNotExist:
-            return JsonResponse({'message':'NO_EXPORTER'}, status=400)
-
-    def delete(self, request, exporter_id):
-        try:
+            exporter_id=request.GET['exporter_id']
             exporter=Exporter.objects.get(id=exporter_id)
             exporter.delete()
             return JsonResponse({'message':'SUCCESS'}, status=200)
 
         except Exporter.DoesNotExist:
             return JsonResponse({'message':'NO_EXPORTER'}, status=400)
+        except KeyError:
+            return JsonResponse({'message':'KEY_ERROR'}, status=400)
